\name{Package.labb3-package}
\alias{Package.labb3-package}
\alias{Package.labb3}
\docType{package}
\title{
\packageTitle{Package.labb3}
}
\description{
\packageDescription{Package.labb3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Package.labb3}
\packageIndices{Package.labb3}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Package.labb3}

Maintainer: \packageMaintainer{Package.labb3}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
