% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.labb3-package.R
\docType{package}
\name{Package.labb3-package}
\alias{Package.labb3}
\alias{Package.labb3-package}
\title{Package.labb3: Euclidian and dijkstra algorithms}
\description{
The packagde has 2 functions, one is an algorithm that calculates the greatest common devider(euclidien algorithm) for two scalars x and y. The other function does an dijkstra algorithm to calculate the shortest distance between a start node and every other node in a given graph. This function takes an data frame with 3 columns named 'v1','v2' and 'w' and a start node(init_node) and returns a vector with the distances.
}
\keyword{internal}
